Split graph visualization into leapp-visualization package
