+++ packages/leapp-visualization/pyproject.toml
@@ -1,0 +1,50 @@
+#
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "leapp-visualization"
+version = "0.5.2"
+description = "Reusable static layered graph visualization for LEAPP and similar pipeline graphs"
+readme = "README.md"
+license = "Apache-2.0"
+authors = [
+    {name = "Frank Lai", email = "frlai@nvidia.com"}
+]
+maintainers = [
+    {name = "Frank Lai", email = "frlai@nvidia.com"}
+]
+keywords = ["graph-visualization", "layout", "svg", "png", "pipeline-graphs"]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Developers",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Scientific/Engineering :: Visualization",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+requires-python = ">=3.11"
+dependencies = [
+    "fast-sugiyama>=0.5.3",
+    "Pillow>=10.0.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/nvidia-isaac/leapp"
+Repository = "https://github.com/nvidia-isaac/leapp"
+Issues = "https://github.com/nvidia-isaac/leapp/issues"
+
+[tool.setuptools]
+include-package-data = true
+zip-safe = false
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["leapp_visualization*"]
