+++ packages/leapp-visualization/src/leapp_visualization/__init__.py
@@ -1,11 +1,5 @@
 from .model import VisualEdge, VisualGraph, VisualNode, VisualPort, VisualTerminal
-
-
-def visualize_graph(*args, **kwargs):
-    from .visualize import visualize_graph as _visualize_graph
-
-    return _visualize_graph(*args, **kwargs)
-
+from .render import render_graph
 
 __all__ = [
     "VisualEdge",
@@ -13,5 +7,5 @@ __all__ = [
     "VisualNode",
     "VisualPort",
     "VisualTerminal",
-    "visualize_graph",
+    "render_graph",
 ]
