+++ packages/leapp-visualization/src/leapp_visualization/png_renderer.py
