+++ packages/leapp-visualization/src/leapp_visualization/svg_renderer.py
